+++ src/vcf/metadata/contig.rs
@@ -7,7 +7,7 @@ use crate::vcf::utils::header_parser::*;
 #[derive(Getters, Clone)]
 pub struct ContigMeta {
     id: String,
-    length: Option<i128>,
+    length: Option<i64>,
     assembly: Option<String>,
     url: Option<Url>,
     md5: Option<String>,
@@ -15,7 +15,7 @@ pub struct ContigMeta {
     taxonomy: Option<String>,
 }
 impl ContigMeta {
-    pub fn new(id: String, length: Option<i128>, assembly: Option<String>, url: Option<Url>, md5: Option<String>, species: Option<String>, taxonomy: Option<String>) -> Self {
+    pub fn new(id: String, length: Option<i64>, assembly: Option<String>, url: Option<Url>, md5: Option<String>, species: Option<String>, taxonomy: Option<String>) -> Self {
         ContigMeta {
             id,
             length,
@@ -38,7 +38,7 @@ impl ContigMeta {
             Some(id) => id.to_string(),
             None => { return Err(anyhow!("ID is not found in contig. {:?}", h)); }
         };
-        let length = h.get("length").map(|s| s.parse::<i128>().unwrap());
+        let length = h.get("length").map(|s| s.parse::<i64>().unwrap());
         let assembly = h.get("assembly").map(|s| s.to_string());
         let url = h.get("url").map(|s| Url::parse(s).unwrap());
         let md5 = h.get("md5").map(|s| s.to_string());
